task14

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -4,7 +4,7 @@
 
 void print_python_list_info(PyObject *p)
 {
-	unsigned int size = PyList_Size(p);
+	long int size = PyList_Size(p);
 	PyListObject *obj = (PyListObject *)p;
 
 	printf("[*] Size of the Python List = %li\n", size);
